requirements.txt
# requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
transformers==4.37.1
diffusers==0.25.0
torch==2.1.2
Pillow==10.2.0
python-dotenv==1.0.0
huggingface-hub==0.20.2
python-multipart==0.0.6
typing-extensions==4.9.0
numpy==1.26.3
tqdm==4.66.1
regex==2023.12.25
pytest==7.4.4         # for testing
httpx==0.26.0         # for testing